Add files via upload
This Website is all about Cardiac Arrest, A person who gets a symptom of Heart attack, this website helps us to identify the problem and help to get appointment to the cardiologist. 
The anatomical heart refers to the physical structure of the heart. It is the organ that helps supply blood and oxygen to all parts of the body. The heart is divided by a partition (or septum) into two halves. The halves are, in turn, divided into four chambers. The heart is situated within the chest cavity and surrounded by a fluid-filled sac called the pericardium.